+++ e2e/__tests__/detectOpenHandles.ts
@@ -5,7 +5,6 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import {onNodeVersions} from '@jest/test-utils';
 import runJest, {runContinuous} from '../runJest';
 
 try {
@@ -81,16 +80,14 @@ it('does not report crypto random data', () => {
   expect(textAfterTest).toBe('');
 });
 
-onNodeVersions('>=12', () => {
-  it('does not report ELD histograms', () => {
-    const {stderr} = runJest('detect-open-handles', [
-      'histogram',
-      '--detectOpenHandles',
-    ]);
-    const textAfterTest = getTextAfterTest(stderr);
+it('does not report ELD histograms', () => {
+  const {stderr} = runJest('detect-open-handles', [
+    'histogram',
+    '--detectOpenHandles',
+  ]);
+  const textAfterTest = getTextAfterTest(stderr);
 
-    expect(textAfterTest).toBe('');
-  });
+  expect(textAfterTest).toBe('');
 });
 
 describe('notify', () => {
@@ -109,17 +106,15 @@ describe('notify', () => {
   });
 });
 
-onNodeVersions('>=12', () => {
-  it('does not report timeouts using unref', () => {
-    // The test here is basically that it exits cleanly without reporting anything (does not need `until`)
-    const {stderr} = runJest('detect-open-handles', [
-      'unref',
-      '--detectOpenHandles',
-    ]);
-    const textAfterTest = getTextAfterTest(stderr);
+it('does not report timeouts using unref', () => {
+  // The test here is basically that it exits cleanly without reporting anything (does not need `until`)
+  const {stderr} = runJest('detect-open-handles', [
+    'unref',
+    '--detectOpenHandles',
+  ]);
+  const textAfterTest = getTextAfterTest(stderr);
 
-    expect(textAfterTest).toBe('');
-  });
+  expect(textAfterTest).toBe('');
 });
 
 it('prints out info about open handlers from inside tests', async () => {