+++ packages/expect-utils/src/types.ts
@@ -3,14 +3,7 @@
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
+ *
  */
 
-'use strict';
-
-const BrowserEnv = require('jest-environment-jsdom');
-
-module.exports = class BrowserEnvWithConditions extends BrowserEnv {
-  exportConditions() {
-    return ['browser'];
-  }
-};
+export type Tester = (a: any, b: any) => boolean | undefined;