+++ packages/jest-runtime/src/index.ts
@@ -5,7 +5,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import * as nativeModule from 'module';
+import nativeModule = require('module');
 import * as path from 'path';
 import {URL, fileURLToPath, pathToFileURL} from 'url';
 import {
@@ -50,7 +50,7 @@ import {formatStackTrace, separateMessageFromStack} from 'jest-message-util';
 import type {MockFunctionMetadata, ModuleMocker} from 'jest-mock';
 import {escapePathForRegex} from 'jest-regex-util';
 import Resolver, {ResolveModuleConfig} from 'jest-resolve';
-import Snapshot = require('jest-snapshot');
+import {EXTENSION as SnapshotExtension} from 'jest-snapshot';
 import {createDirectory, deepCyclicCopy} from 'jest-util';
 import {
   createOutsideJestVmPath,
@@ -373,7 +373,7 @@ export default class Runtime {
       console: options?.console,
       dependencyExtractor: config.dependencyExtractor,
       enableSymlinks: config.haste.enableSymlinks,
-      extensions: [Snapshot.EXTENSION].concat(config.moduleFileExtensions),
+      extensions: [SnapshotExtension].concat(config.moduleFileExtensions),
       forceNodeFilesystemAPI: config.haste.forceNodeFilesystemAPI,
       hasteImplModulePath: config.haste.hasteImplModulePath,
       hasteMapModulePath: config.haste.hasteMapModulePath,
@@ -512,7 +512,7 @@ export default class Runtime {
         this._esmoduleRegistry.set(cacheKey, module);
 
         transformResolve();
-      } catch (error: unknown) {
+      } catch (error) {
         transformReject(error);
         throw error;
       }
@@ -838,7 +838,7 @@ export default class Runtime {
         options,
         moduleRegistry,
       );
-    } catch (error: unknown) {
+    } catch (error) {
       moduleRegistry.delete(modulePath);
       throw error;
     }
@@ -1013,7 +1013,7 @@ export default class Runtime {
       } else {
         return this.requireModule<T>(from, moduleName);
       }
-    } catch (e: unknown) {
+    } catch (e) {
       const moduleNotFound = Resolver.tryCastModuleNotFoundError(e);
       if (moduleNotFound) {
         if (
@@ -1279,7 +1279,7 @@ export default class Runtime {
       return this._resolveModule(from, moduleName, {
         conditions: this.cjsConditions,
       });
-    } catch (err: unknown) {
+    } catch (err) {
       const module = this._resolver.getMockModule(from, moduleName);
 
       if (module) {
@@ -1718,7 +1718,7 @@ export default class Runtime {
     let modulePath;
     try {
       modulePath = this._resolveModule(from, moduleName, options);
-    } catch (e: unknown) {
+    } catch (e) {
       const manualMock = this._resolver.getMockModule(from, moduleName);
       if (manualMock) {
         this._shouldMockModuleCache.set(moduleID, true);